stage2: Replace sudoers patch with sudoers.d file
The sudoers.d/group_sudo file requires that the pi user be added to the
sudo group, which is also done.  This allows the same sudo mod to work
for Raspbian and Debian systems.  Fewer things that can accidentally
break.

This patch is being offered upstream because it avoids patching a
conffile the user may someday be asked about preserving "their" changes
to, see RPi-Distro/pi-gen#18 for details of that.