stage3: Replace xpdf with evince-gtk
The xpdf package has been around forever and is the standard and widely
regarded as the lightest weight PDF reader for a Linux desktop.  There's
just one thing: The Raspberry Pi is _not_ really a Linux desktop, even
if it plays at being one.  You'll find (I did) that xpdf performs quite
poorly as a PDF reader on the Pi for any but the smallest PDF files.

The heavier program weight (and certainly heavier dependencies list)
evince-gtk might use more RAM before it opens a single document, but it
can handle hundreds of pages documents with ease on the Raspberry Pi 2.
On the Pi 3, it's a pleasure to use.  (Nothing GUI is pleasant on the
original Pi A/B.  Use pdftotext and read it with less!)