export-noobs: Build Raspple II instead of Raspbian
The export-noobs dir was hardcoded for Raspbian.  We haven't changed
that, but we have restructured it for Raspple II: Our metadata, our OS
name, our icon.

Also, rather than plunking the slides in a tarball in the deployment
directory, we'll use tar as a convenient way to copy a tree.  You could
(and we possibly should) use cp -r for this, but we didn't.  This is
something we'll address later on.  Honestly the whole export-noobs
process needs a rewrite.