export-noobs: Create undated symlink to NOOBS dir
The export-noobs tool doesn't actually build the NOOBS zipball for us,
just populates a dated version directory that should be renamed and
added to it.  Until we decide how to properly handle that going forward,
let's at least create a symlink that can be referenced in a script that
will always grab the latest version and copy it over.

I've already written a rough script to do this to populate an SD card
from an already-downloaded NOOBS-lite release.  I needed this tool for
less headache in testing, but it's got the fundamental bits.  It will be
easy to clean that up once I'm ready for a beta.

We're close now.